+++ packages/assets-controllers/src/assetsUtil.test.ts
@@ -587,20 +587,13 @@ describe('assetsUtil', () => {
     it('should return empty object when chainId not supported', async () => {
       const testTokenAddress = '0x7BEF710a5759d197EC0Bf621c3Df802C2D60D848';
       const mockPriceService = createMockPriceService();
-      const testTokenRatesState: TokenRatesControllerState = {
-        marketData: {},
-        supportedChainIds: {
-          timestamp: Date.now(),
-          data: ['0x1'],
-        },
-      };
 
       const result = await assetsUtil.fetchTokenContractExchangeRates({
         tokenPricesService: mockPriceService,
         nativeCurrency: 'ETH',
         tokenAddresses: [testTokenAddress],
         chainId: '0x0',
-        tokenRatesState: testTokenRatesState,
+        supportedChainIds: ['0x1'],
       });
 
       expect(result).toStrictEqual({});
@@ -612,20 +605,13 @@ describe('assetsUtil', () => {
       jest
         .spyOn(mockPriceService, 'validateCurrencySupported')
         .mockReturnValue(false);
-      const testTokenRatesState: TokenRatesControllerState = {
-        marketData: {},
-        supportedChainIds: {
-          timestamp: Date.now(),
-          data: ['0x1'],
-        },
-      };
 
       const result = await assetsUtil.fetchTokenContractExchangeRates({
         tokenPricesService: mockPriceService,
         nativeCurrency: 'X',
         tokenAddresses: [testTokenAddress],
         chainId: '0x1',
-        tokenRatesState: testTokenRatesState,
+        supportedChainIds: ['0x1'],
       });
 
       expect(result).toStrictEqual({});
@@ -661,20 +647,13 @@ describe('assetsUtil', () => {
           pricePercentChange1d: 100,
         },
       });
-      const testTokenRatesState: TokenRatesControllerState = {
-        marketData: {},
-        supportedChainIds: {
-          timestamp: Date.now(),
-          data: [testChainId],
-        },
-      };
 
       const result = await assetsUtil.fetchTokenContractExchangeRates({
         tokenPricesService: mockPriceService,
         nativeCurrency: testNativeCurrency,
         tokenAddresses: [testTokenAddress],
         chainId: testChainId,
-        tokenRatesState: testTokenRatesState,
+        supportedChainIds: [testChainId],
       });
 
       expect(result).toMatchObject({
@@ -695,20 +674,13 @@ describe('assetsUtil', () => {
         mockPriceService,
         'fetchTokenPrices',
       );
-      const testTokenRatesState: TokenRatesControllerState = {
-        marketData: {},
-        supportedChainIds: {
-          timestamp: Date.now(),
-          data: [testChainId],
-        },
-      };
 
       await assetsUtil.fetchTokenContractExchangeRates({
         tokenPricesService: mockPriceService,
         nativeCurrency: testNativeCurrency,
         tokenAddresses: tokenAddresses as Hex[],
         chainId: testChainId,
-        tokenRatesState: testTokenRatesState,
+        supportedChainIds: [testChainId],
       });
 
       const numBatches = Math.ceil(
@@ -745,19 +717,12 @@ describe('assetsUtil', () => {
         'fetchTokenPrices',
       );
 
-      const testTokenRatesState: TokenRatesControllerState = {
-        marketData: {},
-        supportedChainIds: {
-          timestamp: Date.now(),
-          data: [testChainId],
-        },
-      };
       await assetsUtil.fetchTokenContractExchangeRates({
         tokenPricesService: mockPriceService,
         nativeCurrency: testNativeCurrency,
         tokenAddresses: tokenAddresses as Hex[],
         chainId: testChainId,
-        tokenRatesState: testTokenRatesState,
+        supportedChainIds: [testChainId],
       });
 
       // Expect batches in ascending order